chore: move BlockExecutorStats
