+++ crates/storage/provider/src/traits/mod.rs
@@ -49,7 +49,7 @@ mod withdrawals;
 pub use withdrawals::WithdrawalsProvider;
 
 mod executor;
-pub use executor::{BlockExecutor, BlockExecutorStats, ExecutorFactory, PrunableBlockExecutor};
+pub use executor::{BlockExecutor, ExecutorFactory, PrunableBlockExecutor};
 
 mod chain;
 pub use chain::{
